Script to turn a directory into a key-value store.
Closes #107

Given a directory, copies its content into an object which can be used as a `KeyValueSource`.

The form of the output key-value store is given by the "output" sub-namespace of the parameters.
See `byte_key_value_sink_from_params` for details.
In the most common case of a zip-backed store, create a sub-namespace like this:
.. code-block:: YAML

    output:
      type: zip
      path: path of zip to write to

Each file in the input directory will become one entry in the output store
with a value equal to the binary content of the file.
The key will by default be the file name (not the full path).
This can be changed by specifying the *key_function* parameter.
Currently the only key function supported is *strip_one_extension*,
which removes the first extension (if present) from the file name before using it as the key.

Because the correct way to handle them is unclear,
if any sub-directories of the input directory are encountered, an error will be raised.
This behavior is subject to change in the future and should not be relied upon.